+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - Removed warnings that dynamic pivot features were in private preview, because
   dynamic pivot is now generally available.
 - Fixed a bug in `session.read.options` where `False` Boolean values were incorrectly parsed as `True` in the generated file format.
+- Fixed a bug that `session.sql("select current_role()")` and `session.get_current_role()` have inconsistent results when executed in an owner's right stored procedure.
 
 
 #### Dependency Updates